Export dummy functions for libselinux_vendor
The following functions are exported in libselinux_vendor, but
simply returns an error:

selinux_android_prop_context_handle
selinux_android_restorecon

Bug: 62886649
Test: BOARD_VNDK_VERSION=current m libselinux_vendor -j
Test: m libselinux_vendor -j
Test: device boots
