+++ libselinux/src/android/android_platform.c
@@ -13,6 +13,16 @@ static const struct selinux_opt seopts_file_rootfs[] = {
     { SELABEL_OPT_PATH, "/nonplat_file_contexts" }
 };
 
+static const struct selinux_opt seopts_prop_split[] = {
+    { SELABEL_OPT_PATH, "/system/etc/selinux/plat_property_contexts" },
+    { SELABEL_OPT_PATH, "/vendor/etc/selinux/nonplat_property_contexts"}
+};
+
+static const struct selinux_opt seopts_prop_rootfs[] = {
+    { SELABEL_OPT_PATH, "/plat_property_contexts" },
+    { SELABEL_OPT_PATH, "/nonplat_property_contexts"}
+};
+
 /*
  * XXX Where should this configuration file be located?
  * Needs to be accessible by zygote and installd when
@@ -132,6 +142,31 @@ struct selabel_handle* selinux_android_file_context_handle(void)
         return selinux_android_file_context(seopts_file_rootfs,
                                             ARRAY_SIZE(seopts_file_rootfs));
     }
+}
+
+struct selabel_handle* selinux_android_prop_context_handle(void)
+{
+    struct selabel_handle* sehandle;
+    const struct selinux_opt* seopts_prop;
+
+    // Prefer files from /system & /vendor, fall back to files from /
+    if (access(seopts_prop_split[0].value, R_OK) != -1) {
+        seopts_prop = seopts_prop_split;
+    } else {
+        seopts_prop = seopts_prop_rootfs;
+    }
+
+    sehandle = selabel_open(SELABEL_CTX_ANDROID_PROP,
+            seopts_prop, 2);
+    if (!sehandle) {
+        selinux_log(SELINUX_ERROR, "%s: Error getting property context handle (%s)\n",
+                __FUNCTION__, strerror(errno));
+        return NULL;
+    }
+    selinux_log(SELINUX_INFO, "SELinux: Loaded property_contexts from %s & %s.\n",
+            seopts_prop[0].value, seopts_prop[1].value);
+
+    return sehandle;
 }
 
 enum levelFrom {
